App system cleanups (#51)
* Add app cli to docker-compose.yml

* Remove app.yml v1

* Add missing import

* More cleanups

* Another missing import

* Add mount for apps

* Remove more
