--- app-standard-v1.json.license
@@ -1,3 +1,0 @@
-SPDX-FileCopyrightText: 2021 Citadel and contributors
-
-SPDX-License-Identifier: GPL-3.0-or-later
