+++ app/lib/composegenerator/v3/networking.py
@@ -7,8 +7,7 @@ from lib.citadelutils import parse_dotenv
 import json
 from os import path
 import random
-from lib.composegenerator.v1.networking import assignIp, assignPort
-
+from lib.composegenerator.shared.networking import assignIp, assignPort
 
 def getMainContainerIndex(app: App):
     if len(app.containers) == 1:
@@ -105,4 +104,3 @@ def configureMainPort(app: AppStage2, nodeRoot: str) -> AppStage3:
     with open(envFile, 'a') as f:
         f.write("{}={}\n".format(portAsEnvVar, app.metadata.port))
     return app
-
